+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+* 0.0.3 (2014-10-17)
+
+ * moved from buzz to guzzle
+ * added revertToDefaultDocumentStorage method
+
 * 0.0.2 (2014-09-05)
 
  * removed validator